% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pclose}
\alias{pclose}
\title{close a connection and remove its object from the environment}
\usage{
pclose(con)
}
\arguments{
  \item{con}{the connection to close}
}
\description{
close a connection and remove its object from the environment
}

