\name{pjoin}
\alias{pjoin}
\title{join a character vector and/or a list by a delimiter, like perl's join it
will append all objects and list elements}
\usage{
pjoin(delim = " ", ...)
}
\arguments{
  \item{delim}{delimiter to join by, defaults to space}

  \item{...}{objects to join}
}
\value{
  a joined character
}
\description{
  join a character vector and/or a list by a delimiter,
  like perl's join it will append all objects and list
  elements
}

